+++ handlers/github_organization_repo_create_v2/test/simple_input.rb
@@ -1,0 +1,15 @@
+{
+  'info' => {
+    'access_token' => '',
+    'org_name' => '',
+    'enable_debug_logging' => 'yes'
+  },
+  'parameters' => {
+    'error_handling' => 'Raise Error',
+    'org_name' => '',
+    'visibility' => 'private',
+    'name' => 'temp',
+    'team_id' => '',
+    'description' => 'Temp repo delete immediately'
+  }
+}