+++ changelog.md
@@ -3,3 +3,11 @@ Github [handlers]
     * Added org_name, output and error message handling. 
 * [github_api_v1] (2020-10-19)
     * Initial commit.
+
+Github [handlers]
+* [github_organization_repo_create_v1] (2021-9-24)
+    * updated auth and fixed bug from last commit
+* [github_personal_repo_create_v1] (2021-9-24)
+    * updated auth
+* [github_organization_repo_create_v2] (2021-9-24)
+    * added team_id parameter